move cdc logic to a separate file to be consistent with other db connectors
